Marconi.Cardano.Indexers.CurrentSyncPointQuery

Description

Gather several indexers to expose a query that can provide you block information about the current last synced block and the chain tip.

Documentation

data CurrentSyncPointQueryIndexer event

Constructors

forall coordinator blockInfoIndexer blockInfoEvent chainTipIndexer chainTipEvent.(IsSync (ExceptT IndexerError IO) event coordinator, IsSync (ExceptT (QueryError (GetLastQuery ChainTip)) IO) chainTipEvent chainTipIndexer, Queryable (ExceptT (QueryError (LatestEventsQuery BlockInfo)) IO) blockInfoEvent (LatestEventsQuery BlockInfo) blockInfoIndexer, Queryable (ExceptT (QueryError (GetLastQuery ChainTip)) IO) chainTipEvent (GetLastQuery ChainTip) chainTipIndexer, Point blockInfoEvent ~ ChainPoint, Point chainTipEvent ~ ChainPoint) => CurrentSyncPointQueryIndexer

Fields

lastPointCoordinator :: coordinator event

blockInfoIndexer :: MVar (blockInfoIndexer blockInfoEvent)

chainTipIndexer :: MVar (chainTipIndexer chainTipEvent)

data CurrentSyncPointQuery

Constructors

CurrentSyncPointQuery

data CurrentSyncPointResult

Constructors

CurrentSyncPointResult

Fields

currentBlockInfo :: BlockInfo

currentTip :: ChainTip