Marconi.Cardano.Indexers.SnapshotBlockEvent

Documentation

snapshotBlockEventWorker :: forall input m n. (MonadIO m, MonadError IndexerError m, MonadIO n) => StandardWorkerConfig n input SnapshotBlockEvent -> SnapshotWorkerConfig input -> FilePath -> m (WorkerIndexer n input SnapshotBlockEvent (WithTrace n (FileIndexer SnapshotMetadata)))

Builds the worker on top of the BlockEvent indexer. This is where the events are preprocessed by filtering out any blocks which are not in the given BlockRange.

newtype SnapshotBlockEvent

Constructors

SnapshotBlockEvent

Fields

getBlockEvent :: BlockEvent

data SnapshotWorkerConfig input

Type which contains the data needed to configure the snapshot indexer workers for BlockEvents and ExtLedgerState events.

Constructors

SnapshotWorkerConfig

Fields

currentBlockNo :: input -> BlockNo

retrieves the current block number from the indexer's input

blockRange :: BlockRange

sub-chain to index, passed by the user on the CL

nodeConfig :: FilePath

configuration of the Cardano node, passed by the user on the CL

data SnapshotMetadata

A representation of the names provided for the snapshot files. This is used by the file indexer when parsing and unparsing the file names.

Constructors

SnapshotMetadata

Fields

snapshotMetadataBlockNo :: Maybe BlockNo

the block which was serialized

snapshotMetadataChainpoint :: ChainPoint

the chain point which was serialized

getConfigCodec :: (MonadIO m, MonadError IndexerError m) => FilePath -> m CodecConfig

For testing

type CodecConfig = CodecConfig (HardForkBlock (CardanoEras StandardCrypto))

type BlockNodeToClientVersion = BlockNodeToClientVersion (CardanoBlock StandardCrypto)

deserialiseSnapshotBlockEvent :: CodecConfig -> BlockNodeToClientVersion -> SnapshotMetadata -> ByteString -> Either Text (Maybe SnapshotBlockEvent)

deserialiseMetadata :: [Text] -> Maybe SnapshotMetadata

blockNodeToNodeVersionM :: Maybe (HardForkNodeToClientVersion (ByronBlock ': CardanoShelleyEras StandardCrypto))